+++ protected/humhub/modules/content/models/forms/MoveContentForm.php
@@ -47,7 +47,8 @@ class MoveContentForm extends Model
     public function rules()
     {
         return [
-            [['target'], 'validateTarget']
+            [['target'], 'validateTarget'],
+            [['target'], 'required']
         ];
     }